+++ src/js/utils/WebColorsContrast.js
@@ -1,0 +1,11 @@
+"use strict";
+/**
+ * @requires Color
+ *
+ * @date 2020-10-27
+ **/
+Object.defineProperty(exports, "__esModule", { value: true });
+var WebColors_1 = require("./WebColors");
+var ORDER_CONTRAST = [8, 0, 6, 4, 1, 9, 2, 10, 3, 7, 5];
+exports.WebColorsContrast = WebColors_1.shuffleWebColors(ORDER_CONTRAST);
+//# sourceMappingURL=WebColorsContrast.js.map