+++ src/js/utils/WebColorsMalachite.d.ts
@@ -1,0 +1,2 @@
+import { Color } from "./datastructures/Color";
+export declare const WebColorsMalachite: Array<Color>;