+++ src/js/utils/WebColorsMalachite.js
@@ -1,0 +1,18 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var Color_1 = require("./datastructures/Color");
+// A mix of green shades
+exports.WebColorsMalachite = [
+    Color_1.Color.makeRGB(0, 21, 6),
+    Color_1.Color.makeRGB(0, 30, 12),
+    Color_1.Color.makeRGB(0, 52, 28),
+    Color_1.Color.makeRGB(0, 81, 47),
+    Color_1.Color.makeRGB(21, 134, 88),
+    Color_1.Color.makeRGB(0, 46, 19),
+    Color_1.Color.makeRGB(0, 68, 40),
+    Color_1.Color.makeRGB(11, 81, 55),
+    Color_1.Color.makeRGB(0, 91, 46),
+    Color_1.Color.makeRGB(0, 111, 46),
+    Color_1.Color.makeRGB(33, 140, 106)
+];
+//# sourceMappingURL=WebColorsMalachite.js.map