+++ src/ts/utils/WebColorsMalachite.ts
@@ -1,0 +1,17 @@
+
+import { Color } from "./datastructures/Color";
+
+// A mix of green shades
+export const WebColorsMalachite : Array<Color> = [
+    Color.makeRGB(0,21,6),  // dark 
+    Color.makeRGB(0,30,12),
+    Color.makeRGB(0,52,28),
+    Color.makeRGB(0,81,47),
+    Color.makeRGB(21,134,88),
+    Color.makeRGB(0,46,19),
+    Color.makeRGB(0,68,40),
+    Color.makeRGB(11,81,55),
+    Color.makeRGB(0,91,46),
+    Color.makeRGB(0,111,46),
+    Color.makeRGB(33,140,106)
+];